align test_snmp_interfaces_error_discard with recent updates
